extract common codes into a class
