+++ src/onscreen.ts
@@ -1,6 +1,7 @@
 import { Config, ErrorIncompatible } from "./Constants";
 import { addStyleIfNotExists } from "./Styles";
-import { addDomIfNotExists, Widget } from "./Widget";
+import { addDomIfNotExists } from "./WidgetBase";
+import { WidgetOnscreen } from "./WidgetOnscreen";
 
 if (!Object.prototype.hasOwnProperty.call(window, "fetch") || !Object.prototype.hasOwnProperty.call(window, "caches")) {
   throw new Error(ErrorIncompatible);
@@ -10,7 +11,7 @@ function companion(options: Config): void {
   addStyleIfNotExists();
   addDomIfNotExists();
 
-  const widget = new Widget(options);
+  const widget = new WidgetOnscreen(options);
   void widget.main();
 }
